[change] Added abstraction to WhoIs related functions
To improve modularity and flexibility, all
WhoIs related code is kept under `whois` subdirectory
under `config` module.
This includes tests, tasks, on delete handlers as well.
Added `cache` to prevent repeated lookups during
multiple WhoIs details fetch during device listing.
