+++ openwisp_controller/config/handlers.py
@@ -153,11 +153,3 @@ def organization_disabled_handler(instance, **kwargs):
         # No change in is_active
         return
     tasks.invalidate_device_checksum_view_cache.delay(str(instance.id))
-
-
-# Remove the related WhoIs record for that ip. If other active devices
-# are linked to it, then new lookup will be triggered for them.
-def device_whois_info_delete_handler(instance, **kwargs):
-    transaction.on_commit(
-        lambda: WhoIsInfo.objects.filter(ip_address=instance.last_ip).delete()
-    )
