+++ openwisp_controller/config/whois/__init__.py
@@ -1,0 +1,2 @@
+# This file is created to allow test runners to discover the tests
+# defined in this directory.
