+++ openwisp_controller/config/whois/handlers.py
@@ -1,0 +1,10 @@
+from django.db import transaction
+
+from .services import WhoIsService
+
+
+# Remove the related WhoIs record for that ip from db and cache
+# If other active devices are linked to it, then new lookup will
+# be triggered for them.
+def device_whois_info_delete_handler(instance, **kwargs):
+    transaction.on_commit(lambda: WhoIsService.delete_whois_record(instance.last_ip))
